ADD: move formatPort to utils;
