+++ src/renderer/utils/api.ts
@@ -5,13 +5,13 @@ import {
   PaginatedQueryParams,
   PaginatedResults,
   PrimaryValidatorConfig,
+  ProtocolType,
   RawBankConfig,
   RawPrimaryValidatorConfig,
 } from '@renderer/types';
 import {formatQueryParams} from '@renderer/utils/address';
 import {SetError, SetResults} from '@renderer/utils/store';
 import {AXIOS_TIMEOUT_MS} from '@renderer/config';
-
 export async function fetchPaginatedResults<T>(
   address: string,
   urlParam: string,
@@ -70,11 +70,29 @@ export const sanitizePortFieldFromRawPrimaryValidatorConfig = (
   data: RawPrimaryValidatorConfig,
   address?: string,
 ): PrimaryValidatorConfig => {
-  // if (address?.includes('https')) {
-  //   return formatAddress('https', address?.replace('https://', ''), undefined);
-  // }
   return {
     ...data,
     port: isInsecureHttp(data.protocol) ? replaceNullPortFieldWithDefaultValue(data.port) : undefined,
   };
+};
+
+
+interface FormatPortArgs {
+  port: string;
+  protocol: ProtocolType;
+}
+
+export const formatPort = ({port, protocol}: FormatPortArgs) => {
+  if (isInsecureHttp(protocol) && port) {
+    return Number(port);
+  }
+
+  if (isInsecureHttp(protocol) && port === undefined) {
+    return 80;
+  }
+
+  const isHttps = !isInsecureHttp(protocol);
+  if (isHttps) {
+    return undefined;
+  }
 };
